+++ packages/primitive/deco-entries/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @spare/deco-entries
 
+## 0.16.10
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/node@0.16.10
+
+## 0.16.9
+
+### Patch Changes
+
+- update interface
+- Updated dependencies
+  - @spare/node@0.16.9
+
 ## 0.16.8
 
 ### Patch Changes